requirements.txt
imgaug==0.3.0
matplotlib==3.2.0
numpy==1.16.0
nvidia-ml-py3==7.352.0
opencv-python==4.1.1.26
Pillow==5.4.1
pydicom==1.2.2
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.3.2
SimpleITK==1.2.0
six==1.12.0
sklearn==0.0
tensorboard==1.13.1
tensorboardX==1.6
tensorflow==1.13.1
tensorflow-estimator==1.13.0
torch==1.3.1
torchvision==0.4.2
tqdm